Documentation grooming
