+++ src/dock_state/tree/node_index.rs
@@ -1,3 +1,5 @@
+use std::ops::Range;
+
 /// Wrapper around indices to the collection of nodes inside a [`Tree`](crate::Tree).
 #[derive(Clone, Copy, PartialEq, Eq, Hash, Debug)]
 #[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize))]
@@ -15,9 +17,9 @@ impl NodeIndex {
     ///
     /// In the context of a [`Tree`](crate::Tree), this will be the node that contains all other nodes.
     ///
-    /// # Example usage
-    /// splitting the current tree in two
+    /// # Examples
     ///
+    /// Splitting the current tree in two.
     /// ```rust
     /// # use egui_dock::{DockState, NodeIndex};
     /// let mut dock_state = DockState::new(vec!["tab 1", "tab 2"]);
@@ -56,20 +58,20 @@ impl NodeIndex {
         (usize::BITS - (self.0 + 1).leading_zeros()) as usize
     }
 
-    /// Returns true if current node is the left node of its parent, false otherwise.
+    /// Returns `true` if current node is the left child of its parent, otherwise `false`.
     #[inline(always)]
     pub const fn is_left(self) -> bool {
         self.0 % 2 != 0
     }
 
-    /// Returns true if current node is the right node of its parent, false otherwise.
+    /// Returns `true` if current node is the right child of its parent, otherwise `false`.
     #[inline(always)]
     pub const fn is_right(self) -> bool {
         self.0 % 2 == 0
     }
 
     #[inline]
-    pub(super) const fn children_at(self, level: usize) -> std::ops::Range<usize> {
+    pub(super) const fn children_at(self, level: usize) -> Range<usize> {
         let base = 1 << level;
         let s = (self.0 + 1) * base - 1;
         let e = (self.0 + 2) * base - 1;
@@ -77,17 +79,17 @@ impl NodeIndex {
     }
 
     #[inline]
-    pub(super) const fn children_left(self, level: usize) -> std::ops::Range<usize> {
+    pub(super) const fn children_left(self, level: usize) -> Range<usize> {
         let base = 1 << level;
         let s = (self.0 + 1) * base - 1;
-        let e = (self.0 + 1) * base + base / 2 - 1;
+        let e = (self.0 + 1) * base + (base / 2) - 1;
         s..e
     }
 
     #[inline]
-    pub(super) const fn children_right(self, level: usize) -> std::ops::Range<usize> {
+    pub(super) const fn children_right(self, level: usize) -> Range<usize> {
         let base = 1 << level;
-        let s = (self.0 + 1) * base + base / 2 - 1;
+        let s = (self.0 + 1) * base + (base / 2) - 1;
         let e = (self.0 + 2) * base - 1;
         s..e
     }
