+++ src/dock_state/window_state.rs
@@ -1,14 +1,14 @@
 use egui::{Pos2, Rect, Vec2};
 
-/// The state for a [`Window`](crate::Surface::Window) surface.
+/// The state of a [`Surface::Window`](crate::Surface::Window) surface.
 ///
-/// Doubles up as a handle for the surface, allowing the user to set its size and position.
+/// Doubles as a handle for the surface, allowing the user to set its size and position.
 #[derive(Debug, Clone)]
 pub struct WindowState {
-    /// The rect which this window last was taking up.
+    /// The [`Rect`] that this window was last taking up.
     screen_rect: Rect,
 
-    /// Was this window dragged last frame?
+    /// Was this window dragged in the last frame?
     dragged: bool,
 
     /// The next position this window should be set to next frame.
@@ -47,13 +47,13 @@ impl WindowState {
         self
     }
 
-    /// Get the rect which this window occupies
-    /// If this window hasn't been shown before, this will be ``Rect::NOTHING``
+    /// Get the [`Rect`] which this window occupies.
+    /// If this window hasn't been shown before, this will be [`Rect::NOTHING`].
     pub fn rect(&self) -> Rect {
         self.screen_rect
     }
 
-    ///
+    /// Returns if this window is currently being dragged or not.
     pub fn dragged(&self) -> bool {
         self.dragged
     }
@@ -82,71 +82,3 @@ impl WindowState {
             })
     }
 }
-
-/*
-impl<Tab> TabWindow {
-
-    ///Destroy the Window and get the Tab back
-    pub fn into_tab(self) -> Tab {
-        self.tab
-    }
-
-    ///Give the window a new location on the Screen
-    pub fn set_position(&mut self, position: Pos2) {
-        self.next_position = Some(position);
-    }
-
-    ///Show the window
-    pub(crate) fn show(
-        &mut self,
-        ui: &mut Ui,
-        tab_viewer: &mut impl TabViewer<Tab = Tab>,
-        window_index: SurfaceIndex,
-        open: Option<&mut bool>,
-    ) -> (Option<Pos2>, Option<TabSource>) {
-        let title = tab_viewer.title(&mut self.tab);
-        let id = Id::new(title.text()).with("TabWindow");
-        let mut window = egui::Window::new(title).id(id);
-        if let Some(position) = self.next_position.take() {
-            window = window.current_pos(position);
-        }
-        if let Some(open) = open {
-            window = window.open(open);
-        }
-
-        let response = window.show(ui.ctx(), |ui| {
-            tab_viewer.ui(ui, &mut self.tab);
-        });
-
-        let screen_rect = {
-            if let Some(response) = response {
-                response.response.rect
-            } else {
-                Rect::NOTHING
-            }
-        };
-
-        //using the min of the screen rect makes sure we don't catch resizings.
-        //using self.dragged as an alternative condition makes sure the drag doesn't stop if we stop moving mid drag.
-        if screen_rect.min != self.screen_rect.min || self.dragged {
-            self.screen_rect = screen_rect;
-            let something_dragged = ui.memory(|mem| mem.is_anything_being_dragged());
-
-            //this enforces the drag start pattern which tabs follow, that is it's Some for the first frame of the drag, then none.
-            let drag_start = if something_dragged && !self.dragged {
-                Some(screen_rect.min)
-            } else {
-                None
-            };
-
-            self.dragged = something_dragged;
-            (
-                drag_start,
-                something_dragged.then_some(TabSource::Window(window_index)),
-            )
-        } else {
-            (None, None)
-        }
-    }
-}
- */
