+++ src/widgets/dock_area/allowed_splits.rs
@@ -1,13 +1,16 @@
-/// What directions can this dock split in?
+/// What directions can this dock be split in?
 #[derive(Copy, Clone, Debug, Default, Eq, PartialEq)]
 pub enum AllowedSplits {
     #[default]
     /// Allow splits in any direction (horizontal and vertical).
     All = 0b11,
-    /// Only allow split in a horizontal direction.
+
+    /// Only allow split in a horizontal directions.
     LeftRightOnly = 0b10,
-    /// Only allow splits in a vertical direction.
+
+    /// Only allow splits in a vertical directions.
     TopBottomOnly = 0b01,
+
     /// Don't allow splits at all.
     None = 0b00,
 }
@@ -21,7 +24,7 @@ impl std::ops::BitAnd for AllowedSplits {
 }
 
 impl AllowedSplits {
-    /// Create an allowed splits from a u8, panics if an invalid value is given.
+    /// Create allowed splits from a u8, panics if an invalid value is given.
     #[inline(always)]
     fn from_u8(u8: u8) -> Self {
         match u8 {
@@ -29,7 +32,7 @@ impl AllowedSplits {
             0b10 => AllowedSplits::LeftRightOnly,
             0b01 => AllowedSplits::TopBottomOnly,
             0b00 => AllowedSplits::None,
-            _ => unreachable!(),
+            _ => panic!("Provided an invalid value for allowed splits: {u8:0x}"),
         }
     }
 }
