+++ src/widgets/tab_viewer.rs
@@ -6,19 +6,19 @@ pub trait TabViewer {
     /// The type of tab in which you can store state to be drawn in your tabs.
     type Tab;
 
+    /// The title to be displayed in the tab bar.
+    fn title(&mut self, tab: &mut Self::Tab) -> WidgetText;
+
     /// Actual tab content.
     fn ui(&mut self, ui: &mut Ui, tab: &mut Self::Tab);
 
-    /// Content inside context_menu.
+    /// Content inside the context menu shown when the tab is right-clicked.
     ///
-    /// The `_node` specifies which `Node` or split of the tree that this
-    /// particular context menu belongs to.
+    /// The `_node` specifies which [`Node`](crate::Node) or split of the tree that this particular context menu
+    /// belongs to.
     fn context_menu(&mut self, _ui: &mut Ui, _tab: &mut Self::Tab, _node: NodeIndex) {}
 
-    /// The title to be displayed.
-    fn title(&mut self, tab: &mut Self::Tab) -> WidgetText;
-
-    /// Unique id for this tab.
+    /// Unique ID for this tab.
     ///
     /// If not implemented, uses tab title text as an id source.
     fn id(&mut self, tab: &mut Self::Tab) -> Id {
@@ -37,28 +37,26 @@ pub trait TabViewer {
 
     /// This is called when the tabs close button is pressed.
     ///
-    /// Returns `true` if the tab should close immediately, `false` otherwise.
+    /// Returns `true` if the tab should close immediately, otherwise `false`.
     ///
-    /// NOTE if returning false `ui` will still be called once more if this tab is active.
+    /// **Note**: if `false` is returned, [`ui`](Self::ui) will still be called once more if this tab is active.
     fn on_close(&mut self, _tab: &mut Self::Tab) -> bool {
         true
     }
 
-    /// This is called when the tabs add button is pressed.
+    /// This is called when the tabs add button is pressed and if the dock [`Style`](crate::Style)'s
+    /// `show_add_buttons` is set to `true`.
     ///
-    /// This requires the dock style's `show_add_buttons` to be `true`.
-    ///
-    /// The `_node` specifies which `Node` or split of the tree that this
-    /// particular add button was pressed on.
+    /// `_node` specifies which [`Node`](crate::Node) or split of the tree this particular add button was pressed on.
     fn on_add(&mut self, _node: NodeIndex) {}
 
-    /// Content of add_popup. Displays a popup under the add button. Useful for selecting
-    /// what type of tab to add.
+    /// Content of the popup under the add button. Useful for selecting what type of tab to add.
     ///
-    /// This requires the dock style's `show_add_buttons` and `show_add_popup` to be `true`.
+    /// This requires that [`DockArea::show_add_buttons`](crate::DockArea::show_add_buttons) and
+    /// [`DockArea::show_add_popup`](crate::DockArea::show_add_popup) are set to `true`.
     fn add_popup(&mut self, _ui: &mut Ui, _node: NodeIndex) {}
 
-    /// This is called every frame after `ui` is called (if the tab is active).
+    /// This is called every frame after [`ui`](Self::ui) is called, if the `_tab` is active.
     ///
     /// Returns `true` if the tab should be forced to close, `false` otherwise.
     ///
@@ -72,9 +70,9 @@ pub trait TabViewer {
         None
     }
 
-    /// Specifies a tabs ability to be shown in a window.
+    /// Specifies a tab's ability to be shown in a window.
     ///
-    /// Returns ``false`` if you don't want this tab to be dragged into or be turned into a window.
+    /// Return `false` if you don't want this tab to be turned into a window.
     fn allowed_in_windows(&self, _tab: &mut Self::Tab) -> bool {
         true
     }
